requirement.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.3.1
django-restframework==0.0.1
docker==7.1.0
gunicorn==22.0.0
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.50
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==308
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0